+++ src/com/excilys/formation/cdb/persistance/dao/SimpleDAO.java
@@ -2,5 +2,5 @@ package com.excilys.formation.cdb.persistance.dao;
 
 public interface SimpleDAO {
 
-    public abstract Integer select(String query);
+    public abstract Long select(String query);
 }
